Adjust margin, padding and borders part 2
This next exercise reiterates using margin, padding and borders in order
to get a desired outcome. I am having a bit of trouble on this one. So far
I think I have the right idea starting out with the html elements in place.

Currently the issue is trying to get the borders to fit right around the
three distinct parts. I am also having issues with the background color,
and padding of the contents inside each box.