Second and third attempts
After reviewing the solution and seeing that my intial code was way too
bulky, I tried this one again keeping in mind to use less lines. There
are things within the CSS and HTML that wasn't really explained to me
that I had to experiment with and Google to figure out.

I decided to use the HTML layout in the solution to see the logic behind
how it was organized. I learned about using "margin: 0 auto" rather
than creating a <p> element for the button, which caused some issues for
me the first time around. Once the HTML layout made more sense to me, I
created a class for the "card" and each section within it.

Most of my time was spent adjusting margin and padding to get the right
results. It still isn't exactly what the solution code provides but it
looks a bit better than yesterday's code.