Flex header (cont. part 3)
It seems like the issue yesterday was trying to get the flex items to
change size with the browser windonw but today the issue is something else.
I noticed a left margin next to <a href="#">ONE</a> and I can't seem to get
rid of it.

I think yesterday's code was a bit closer to the solution if it didn't have
some extra lines like the "left-side" and "right-side" divs.