Flex header 2
This assignment originally asked me to edit the provided HTML file but I
tried constructing it on my own. My approach didn't use an <ul> and <li>
elements and it also included a button rather than a link for the
notification.

I will try this assignment again using the provided HTML and see how I
approach the CSS differently.