Flex header 2 (continued)
This time around I used the provided HTML to complete this assignment. It
was easier in some ways but still difficult in the same ways. My end code
still looks a lot busier compared to the solution. In the end I had to look
at the solution CSS again to see what errors I was making and how to
simplify some areas of my work.

I realize that I am using certain properties multiple times whereas the
solution reveals that only a couple can get the job done. It's still a
little unclear when and how to use these properties this way.