Flex information
This assignment was a bit easier. I used the provided HTML rather than try
to construct everything on my own. It still required me to edit and add
some lines to get it in a good place to work with.

Some of the things that I did that the solution verified was firstly to
create a container for the four plants and their texts. However, the
solution created one class for all the items ("info") but I made one for
each plant on its own and used a grouping selector, which in theory I
guess it does the same thing. Another difference that I noticed was how my
code had containers in a few places but the solution only used one on the
"info" container.