Flex-modal
This assignment had some areas of difficulty but I managed to find a
solution, though it is not 100% like the solution.

Mainly I knew to make a container for the "middle-section" which contained
the "header", "text", and "continue/cancel" buttons. However, the solution
did not create a container for the buttons like I did. I did this because
I was unable to get the buttons to line up horizontally so I decided to
change the flex direction to get them aligned.

I also noticed that the solution placed the "close-button" as a flex item
within the "header" (as a container) to get them lined up. I did not do
this because my approach made them line up, more or less.

Overall, the solution CSS looked a bit longer which I wasn't expecting
because it is usually concise and simpler than mine.