Flex-layout-2
This assignment was surprisingly not too difficult once I figured out the
approach. I knew that each major part of the layout needed to be a
container so the "header", "content", "cards", and "footer" were made to
house their respective divs.

The main challenge was finding the correct properties to style all the
elements. The solution was also very straight-forward and not too far off
from my code.